requirements.txt
requests==2.25.1
discord.py==1.7.2
Pillow==8.2.0
PyNaCl==1.4.0
